refractor: generate functions -> generateUtil
