+++ frontend/src/components/Feature/Diplomadataform/DiplomaDataForm.tsx
@@ -3,7 +3,7 @@ import { useForm } from "react-hook-form";
 import { TemplateResponse, SaltData, Student, FormDataUpdateRequest, TrackResponse, BootcampResponse, pdfGenerationResponse } from "../../../util/types";
 import { FileUpload } from "../../MenuItems/Inputs/FileUploader";
 import { ParseFileData } from '../../../services/InputFileService';
-import { generateVerificationCode, mapBootcampToSaltData2 } from "../../../util/helper";
+import { mapBootcampToSaltData2 } from "../../../util/helper";
 import './DiplomaDataForm.css';
 import { AlertPopup, CustomAlertPopupProps, PopupType } from "../../MenuItems/Popups/AlertPopup";
 import { Template } from "@pdfme/common";
@@ -19,6 +19,7 @@ import { SaveButton } from "../../MenuItems/Buttons/SaveButton";
 import { useCustomAlert } from "../../Hooks/useCustomAlert";
 import { newGenerateAndDownloadZippedPDFs, newGenerateAndPrintCombinedPDF, newGenerateCombinedPDF } from "../../../util/pdfGenerationUtil";
 import { openPrintWindowfromBlob, openWindowfromBlob } from "../../../util/fileActionUtil";
+import { generateVerificationCode } from "../../../util/generateUtil";
 
 type FormData = {
   optionA: boolean;
